+++ packages/event-core/src/eventa.ts
@@ -1,5 +1,3 @@
-import type { Adapter } from './adapters/browser'
-
 import { customAlphabet } from 'nanoid'
 
 interface EventPayload<T> {
@@ -54,82 +52,3 @@ export function defineInvokeEvent<Req, Res>(tag?: string) {
 }
 
 export type InvokeEvent<Req, Res> = ReturnType<typeof defineInvokeEvent<Req, Res>>
-
-interface CreateContextProps {
-  adapter?: Adapter
-
-  // hooks?: {
-  //   onReceive?: (event: Event<any, any>) => void
-  // }
-}
-
-export function createContext(props: CreateContextProps = {}) {
-  const listeners = new Map<EventTag<any, any>, Set<(params: any) => any>>()
-  const onceListeners = new Map<EventTag<any, any>, Set<(params: any) => any>>()
-
-  function emit<Req, Res>(event: EventTag<Req, Res>, payload: Req) {
-    for (const listener of listeners.get(event) || []) {
-      listener(payload)
-    }
-
-    for (const onceListener of onceListeners.get(event) || []) {
-      onceListener(payload)
-      onceListeners.get(event)?.delete(onceListener)
-    }
-  }
-
-  const hooks = props.adapter?.(emit).hooks
-
-  return {
-    // listeners,
-    // onceListeners,
-
-    emit,
-
-    on<Req, Res>(event: EventTag<Req, Res>, handler: (payload: Req) => void) {
-      if (!listeners.has(event)) {
-        listeners.set(event, new Set())
-      }
-      listeners.get(event)?.add((payload: Req) => {
-        handler(payload)
-        hooks?.onReceive?.(event, payload)
-      })
-    },
-
-    once<Req, Res>(event: EventTag<Req, Res>, handler: (payload: Req) => void) {
-      if (!onceListeners.has(event)) {
-        onceListeners.set(event, new Set())
-      }
-      onceListeners.get(event)?.add((payload: Req) => {
-        handler(payload)
-        hooks?.onReceive?.(event, payload)
-      })
-    },
-
-    off<Req, Res>(event: EventTag<Req, Res>) {
-      listeners.delete(event)
-      onceListeners.delete(event)
-    },
-
-    until<Req, Res>(event: EventTag<Req, Res>, listener: (payload: Req) => any): Promise<Res> {
-      return new Promise((resolve, reject) => {
-        if (!onceListeners.has(event)) {
-          onceListeners.set(event, new Set())
-        }
-
-        onceListeners.get(event)?.add((data) => {
-          try {
-            const result = listener(data)
-            resolve(result)
-          }
-          catch (error) {
-            reject(error)
-          }
-        })
-      })
-    },
-  }
-}
-
-export type EventContext = ReturnType<typeof createContext>
-export type EventContextEmitFn = EventContext['emit']
