+++ packages/event-core/src/stream.ts
@@ -1,4 +1,5 @@
-import type { EventContext, InvokeEvent } from './eventa'
+import type { EventContext } from './context'
+import type { InvokeEvent } from './eventa'
 
 export function defineStreamInvoke<Req, Res>(clientCtx: EventContext, event: InvokeEvent<Req, Res>) {
   return (req: Req) => {
